fix all-day event rendering when maxTime. fixes #4483
